change mapping implementation
