+++ lib/features/home/data/responses/user_response.dart
@@ -1,6 +1,9 @@
 import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
+import '../../../../core/utils/constants.dart';
+import '../../domain/entities/user_detail.dart';
+
 part 'user_response.g.dart';
 
 @JsonSerializable()
@@ -19,6 +22,12 @@ class UserResponse extends Equatable {
   });
 
   factory UserResponse.fromJson(Map<String, dynamic> json) => _$UserResponseFromJson(json);
+
+  UserDetail get toDomain => UserDetail(
+    id: id ?? Constants.emptyString,
+    name: name ?? Constants.emptyString,
+    balance: balance ?? 0.0,
+  );
 
   @override
   List<Object?> get props => [id, name, balance];
