+++ test/features/home/data/responses/user_response_test.dart
@@ -2,6 +2,7 @@
 import 'dart:convert';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:wallet/features/home/data/responses/user_response.dart';
+import 'package:wallet/features/home/domain/entities/user_detail.dart';
 
 import '../../../../fixtures/stub_json/fixtures.dart';
 import '../../../../fixtures/stub_json/json_reader.dart';
@@ -13,6 +14,15 @@ void main() {
       const expectedResult = UserResponse(id: "1", name: "test", balance: 1.0);
       final Map<String, dynamic> jsonMap = json.decode(readJson(Fixtures.userResponse));
       final result = UserResponse.fromJson(jsonMap);
+      expect(result, expectedResult);
+    });
+  });
+
+  group("toDomain", () {
+    test("should be mapped properly to UserDetail entity", () async {
+      const response = UserResponse(id: "1", name: "test", balance: 1);
+      const expectedResult = UserDetail(id: "1", name: "test", balance: 1);
+      final result = response.toDomain;
       expect(result, expectedResult);
     });
   });
